Live: implement fix-optimize, add a parameter to LiveExpr
The extra parameter allows us to iterate analyse+dbe.
Previously the types of dead binding elimination were as follows:

Expr Γ
  |
  | (analyse)
  v
Σ[ Δ ∈ Subset Γ ] LiveExpr Γ Δ
  |
  | (dbe)
  v
Σ[ Δ ∈ Subset Γ ] Expr ⌊ Δ ⌋

The semantics of input and output expression can be related using
evaluation with an (env : Env Γ) and (prjEnv Δ env : Env ⌊ Δ ⌋).
However, iterating this process leads to a problem. The Expr ⌊ Δ ⌋ will
be turned into a Σ[ Δ' ∈ Subset ⌊ Δ ⌋ ] Expr ⌊ Δ' ⌋ and then
Σ[ Δ'' ∈ Subset ⌊ Δ' ⌋ ] Expr ⌊ Δ'' ⌋ etc.
The Ctx of the result is related to the original through an unknown
number of Subset indirections (requiring an unknown number of prjEnv
calls for the evaluation environment).

The distinction between Γ and ⌊ Δ ⌋ already caused complications in
other places, too. It seems desirable to work with Expr ⌊ Δ ⌋
everywhere, going to another Expr ⌊ Δ' ⌋ at each iteration, with both
(Δ Δ' : Subset Γ).

Another way of achieving this would have been to pre-compose some
(embed : Expr ⌊ Δ ⌋ → Expr Γ) to make the types line up. It could use
renameExpr and then just go from Expr ⌊ all Γ ⌋ to Expr Γ.
However, this did very artificial to me. The function basically
does nothing and the we would still need to prove it does nothing.

So I decided to directly take an Expr ⌊ Δ ⌋ in analyse.
I still wanted the LiveExpr the Expr to represent where it came from,
i.e. refer to the Δ and store the unmodified references. Without that,
e.g. (analyse : Expr ⌊ Δ ⌋ → Σ[ Δ' ∈ Subset Γ ] LiveExpr Γ Δ') the
trivial (forget : LiveExpr Γ Δ' → Expr Γ) is not its inverse anymore and
we need to rewrite the Refs somewhere.
I addressed this by parametrizing LiveExpr over the Δ it came from:
(analyse : Expr ⌊ Δ ⌋ → Σ[ Δ' ∈ Subset Γ ] LiveExpr Δ Δ') with
(Var : (i : Ref σ ⌊ Δ ⌋) → LiveExpr Δ (Δ [ i ]) σ).
This unfortunately adds some complexity as well in the form of more type
parameters and an additional (H : Δ' ⊆ Δ) that is required in some
places.